修改注释，更改generate_inc_val位置
