+++ service/core.py
@@ -15,17 +15,49 @@ import numpy
 import docs.conf as CONFIG
 from service import helpers
 
-async def update_min_collection(handler, record_bson, record_bson_old={}):
-    """更新[合并数据]的v1, v2, v3数值
-    
-    将相似的[原始数据]合并至同一分钟级，
 
-    根据更新后的[合并数据]和更新后与更行前的差值生成符合$inc格式的dict，
-    并更新[合并数据]的归一值
+def generate_v_val_inc_query(record_bson, record_bson_old={}):
+    """生成符合find_one_and_update中$inc要求格式的v1，v2，v3值
+
+    将dict格式的v3值从v3 : {attr : val}转化为v3.attr : val
+    在结果dict中插入v1，v2，修改格式后的v3并返回
+
+    参数：
+        record_bson (dict)：需要合并的[原始数据]，数据为BSON类型
+        record_bson_old (dict)：更新[原始数据]的情况下[原始数据]的原值，数据为BSON类型
+
+    返回：
+        dict：修改格式后的v1，v2，v3值
+
+    """
+    result = {}
+    result_dict = {}
+    result['v1'] = record_bson['v1'] - (record_bson_old['v1'] if record_bson_old else 0)
+    result['v2'] = record_bson['v2'] - (record_bson_old['v2'] if record_bson_old else 0)
+    result_dict['v1'] = result['v1']
+    result_dict['v2'] = result['v2']
+    result_dict['v3'] = {}
+    for key in record_bson['v3'].keys():
+        result[''.join(['v3.', key])] = record_bson['v3'][key] - (record_bson_old['v3'][key] if record_bson_old and key in record_bson_old['v3'] else 0)
+        result_dict['v3'][key] = result[''.join(['v3.', key])]
+    if record_bson_old:
+        for key in record_bson_old['v3'].keys():
+            if key not in record_bson['v3'].keys():
+                result[''.join(['v3.', key])] = -record_bson_old['v3'][key]
+                result_dict['v3'][key] = result[''.join(['v3.', key])]
+    return result, result_dict
+
+async def update_min_collection(handler, record_bson, record_bson_old={}):
+    """更新[合并数据]的v1, v2, v3数值，将相似的[原始数据]合并至同一分钟级
+    
+    根据[原始数据]原值与新值生成符合$inc格式的v1，v2，v3增值dict
+    *注：当原值存在新值没有的键时，更新[合并数据]时，减去该键值和键值的归一值
+    更新[合并数据]除归一值外的值(第二次数据库操作)
 
     参数：
         handler (tornado.web.RequestHandler)：Tornado的HTTP Request Handler
         record_bson (dict)：需要合并的[原始数据]，数据为BSON类型
+        record_bson_old (dict)：更新[原始数据]的情况下[原始数据]的原值，数据为BSON类型
 
     """
     # 计算合并时间范围下限
@@ -38,9 +70,10 @@ async def update_min_collection(handler, record_bson, record_bson_old={}):
     datetime_end = datetime_begin + datetime.timedelta(minutes=1)
     # 符合$inc所需格式的v1, v2, v3增值
     if record_bson_old:
-        inc_val, inc_val_dict = helpers.generate_v_val_inc_query(record_bson, record_bson_old=record_bson_old)
+        inc_val, inc_val_dict = generate_v_val_inc_query(record_bson, record_bson_old=record_bson_old)
     else:
-        inc_val, inc_val_dict = helpers.generate_v_val_inc_query(record_bson)
+        inc_val, inc_val_dict = generate_v_val_inc_query(record_bson)
+    # 第二次数据库操作
     # 根据条件寻找符合条件的[合并数据]并更新，可以根据情况增减条件(会影响[合并数据]集合大小)
     after_update_data = await handler.settings['db'][CONFIG.MIN_COLLECTION_NAME] \
             .find_one_and_update( \
@@ -74,9 +107,8 @@ async def update_min_collection(handler, record_bson, record_bson_old={}):
 async def update_min_collection_norm_val(handler, new_record, inc):
     """更新[合并数据]的归一值
     
-    根据更新后的[合并数据]和更新后与更行前的差值生成符合$inc格式的dict，
-    如果[合并数据]更新前归一值为0则计算归一值，否则计算归一值增值，
-    并更新[合并数据]的归一值
+    根据[合并数据]原值与新值生成符合$inc格式的归一值增值dict
+    更新[合并数据]的归一值(第三次数据库操作)
 
     参数：
         handler (tornado.web.RequestHandler)：Tornado的HTTP Request Handler
@@ -93,6 +125,7 @@ async def update_min_collection_norm_val(handler, new_record, inc):
             # 修改v3归一值/归一值增值格式
             inc_params['v3_norm.'+str(key)] = helpers.log10_addition_normalize( \
                     new_record['v3'][key] - inc['v3'][key], inc['v3'][key])
+    # 第三次数据库操作
     # 更新归一值
     after_update_data = await handler.settings['db'][CONFIG.MIN_COLLECTION_NAME] \
             .find_one_and_update(
